Upgrade playwright, MCR and node version
- For versions of Playwright >= 1.46.0 the code
coverage metrics reported for classes with instance
variables changed (it became more precise). So, we
needed to upgrade the version of node we use to
23.10.0 where the same change is available there so
that when code coverage is generated on the node side
it will match the ranges produced by the browser for
the same module. When these don't match we were
seeing mistaken coverage reports -- code that was
clearly executed was reported as uncovered. And this
caused a failing test, where we check that code
coverage from node and the browser is properly merged
- Upgraded MCR as well, but this is independent
of making the test pass again.